Add Nats reciever and sender
