+++ common.go
@@ -25,9 +25,11 @@
 package skogul
 
 import (
+	"os"
 	"fmt"
 	"math"
 	"runtime"
+        "crypto/x509"
 
 	"github.com/sirupsen/logrus"
 )
@@ -381,3 +383,30 @@ func IsInf(f float32, sign int) bool {
 func MissingArgument(field string) error {
 	return fmt.Errorf("missing required configuration option `%s'", field)
 }
+
+//Make GetCertPool a common function instead of reciever specific.
+func GetCertPool(path string) (*x509.CertPool, error) {
+        // this means "use system default"
+        if path == "" {
+                return nil, nil
+        }
+        cp := x509.NewCertPool()
+        fd, err := os.Open(path)
+        if err != nil {
+                return nil, fmt.Errorf("unable to open custom root CA: %w", err)
+        }
+        defer func() {
+                fd.Close()
+        }()
+        bytes := make([]byte, 1024000)
+        n, err := fd.Read(bytes)
+        if err != nil {
+                return nil, fmt.Errorf("unable to read custom root CA: %w", err)
+        }
+        ok := cp.AppendCertsFromPEM(bytes[:n])
+        if !ok {
+                return nil, fmt.Errorf("unable to append certificate to root CA pool")
+        }
+        return cp, nil
+}
+
